+++ demo/src/main/java/com/orange/ods/demo/ui/guidelines/color/GuidelineColorScreen.kt
@@ -57,7 +57,7 @@ import com.orange.ods.demo.ui.LocalMainTopAppBarManager
 import com.orange.ods.demo.ui.LocalOdsDemoGuideline
 import com.orange.ods.demo.ui.utilities.composable.Title
 import com.orange.ods.demo.ui.utilities.getStringName
-import com.orange.ods.theme.OdsColors
+import com.orange.ods.theme.colors.OdsColors
 import com.orange.ods.theme.guideline.GuidelineColor
 import com.orange.ods.theme.guideline.GuidelineColorType
 import com.orange.ods.theme.guideline.toHexString