+++ demo/src/main/java/com/orange/ods/demo/ui/utilities/extension/WebViewExt.kt
@@ -14,7 +14,7 @@ import android.webkit.WebView
 import androidx.annotation.RawRes
 import androidx.compose.ui.graphics.Color
 import com.orange.ods.demo.R
-import com.orange.ods.theme.OdsColors
+import com.orange.ods.theme.colors.OdsColors
 import com.orange.ods.theme.guideline.toHexString
 
 fun WebView.injectLightDarkModeCss(isDarkModeEnabled: Boolean, colors: OdsColors) {