+++ lib/src/main/java/com/orange/ods/compose/component/bottomnavigation/OdsBottomNavigation.kt
@@ -53,8 +53,8 @@ fun OdsBottomNavigation(
     BottomNavigation(
         modifier = modifier,
         // Need to define backgroundColor cause in Compose default backgroundColor is primarySurface
-        backgroundColor = OdsTheme.colors.bottomNavigation.barBackground,
-        contentColor = OdsTheme.colors.bottomNavigation.barContent,
+        backgroundColor = OdsTheme.colors.component.bottomNavigation.barBackground,
+        contentColor = OdsTheme.colors.component.bottomNavigation.barContent,
         content = content
     )
 }
@@ -112,8 +112,8 @@ fun RowScope.OdsBottomNavigationItem(
             }
         },
         alwaysShowLabel = alwaysShowLabel,
-        selectedContentColor = OdsTheme.colors.bottomNavigation.itemSelected,
-        unselectedContentColor = OdsTheme.colors.bottomNavigation.itemUnselected
+        selectedContentColor = OdsTheme.colors.component.bottomNavigation.itemSelected,
+        unselectedContentColor = OdsTheme.colors.component.bottomNavigation.itemUnselected
     )
 }